Split Factories into several fiels
