+++ members/tests/factories/__init__.py
@@ -1,0 +1,25 @@
+from .address_factory import AddressFactory
+from .union_factory import UnionFactory
+from .department_factory import DepartmentFactory
+from .family_factory import FamilyFactory
+from .person_factory import PersonFactory
+from .activity_factory import ActivityFactory
+from .member_factory import MemberFactory
+from .waitinglist_factory import WaitingListFactory
+from .activity_participant_factory import ActivityParticipantFactory
+from .payment_factory import PaymentFactory
+from .volunteer_factory import VolunteerFactory
+
+__all__ = [
+    ActivityParticipantFactory,
+    PaymentFactory,
+    AddressFactory,
+    UnionFactory,
+    WaitingListFactory,
+    DepartmentFactory,
+    FamilyFactory,
+    PersonFactory,
+    ActivityFactory,
+    MemberFactory,
+    VolunteerFactory,
+]
