+++ members/tests/factories/factory_helpers.py
@@ -1,0 +1,19 @@
+import pytz
+import random
+from datetime import timedelta
+
+from django.conf import settings
+
+
+LOCALE = "dk_DK"
+TIMEZONE = pytz.timezone(settings.TIME_ZONE)
+# Setting default locale (this is not documented or intended by factory_boy)
+# Faker._DEFAULT_LOCALE = LOCALE
+
+
+def datetime_before(datetime):
+    return datetime - timedelta(days=random.randint(1, 4 * 365))
+
+
+def datetime_after(datetime):
+    return datetime + timedelta(days=random.randint(1, 4 * 365))
