+++ members/tests/factories/volunteer_factory.py
@@ -1,0 +1,19 @@
+from members.tests.factories.factory_helpers import TIMEZONE
+from members.models import Volunteer
+from members.tests.factories.person_factory import PersonFactory
+from members.tests.factories.department_factory import DepartmentFactory
+from factory import Faker, DjangoModelFactory, SubFactory, LazyAttribute
+
+from members.tests.factories.factory_helpers import datetime_after
+
+
+class VolunteerFactory(DjangoModelFactory):
+    class Meta:
+        model = Volunteer
+
+    person = SubFactory(PersonFactory)
+    department = SubFactory(DepartmentFactory)
+
+    added = Faker("date_time", tzinfo=TIMEZONE)
+    confirmed = LazyAttribute(lambda d: datetime_after(d.added))
+    removed = LazyAttribute(lambda d: datetime_after(d.added))
